Add location filtering functionality for entire app
